+++ internal/command/upload_test.go
@@ -41,6 +41,7 @@ func TestUploadConfigurationCommandRun(t *testing.T) {
 		Workspace   string
 		Directory   string
 		Speculative bool
+		Provisional bool
 	}
 
 	type args struct {
@@ -62,6 +63,7 @@ func TestUploadConfigurationCommandRun(t *testing.T) {
 				Workspace:   "ws-1",
 				Directory:   "dir/",
 				Speculative: false,
+				Provisional: false,
 			},
 			want: 0,
 		},
@@ -74,6 +76,7 @@ func TestUploadConfigurationCommandRun(t *testing.T) {
 				Workspace:   tt.fields.Workspace,
 				Directory:   tt.fields.Directory,
 				Speculative: tt.fields.Speculative,
+				Provisional: tt.fields.Provisional,
 			}
 			if got := c.Run(tt.args.args); got != tt.want {
 				t.Errorf("Run() = %v, want %v", got, tt.want)