+++ internal/command/run_create.go
@@ -20,8 +20,9 @@ type CreateRunCommand struct {
 	ConfigurationVersionID string
 	Message                string
 
-	PlanOnly bool
-	SavePlan bool
+	PlanOnly  bool
+	IsDestroy bool
+	SavePlan  bool
 }
 
 func (c *CreateRunCommand) flags() *flag.FlagSet {
@@ -30,6 +31,7 @@ func (c *CreateRunCommand) flags() *flag.FlagSet {
 	f.StringVar(&c.ConfigurationVersionID, "configuration_version", "", "The Configuration Version ID to use for this run.")
 	f.StringVar(&c.Message, "message", "", "Specifies the message to be associated with this run. A default message will be set.")
 	f.BoolVar(&c.PlanOnly, "plan-only", false, "Specifies if this is a Terraform Cloud speculative, plan-only run that cannot be applied.")
+	f.BoolVar(&c.IsDestroy, "is-destroy", false, "Specifies that the plan is a destroy plan. When true, the plan destroys all provisioned resources.")
 	f.BoolVar(&c.SavePlan, "save-plan", false, "Specifies whether to create a saved plan. Saved-plan runs perform their plan and checks immediately, but won't lock the workspace and become its current run until they are confirmed for apply.")
 	return f
 }
@@ -52,6 +54,7 @@ func (c *CreateRunCommand) Run(args []string) int {
 		ConfigurationVersionID: c.ConfigurationVersionID,
 		Message:                c.Message,
 		PlanOnly:               c.PlanOnly,
+		IsDestroy:              c.IsDestroy,
 		SavePlan:               c.SavePlan,
 		RunVariables:           runVars,
 	})