+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ## Enhancements
 * Adds new command, `workspace output list` to retrieve outputs for an existing Terraform Cloud workspace by @mjyocca [#29](https://github.com/hashicorp/tfc-workflows-tooling/pull/29)
 
+* Add support for saved plans by @1newsr [#57](https://github.com/hashicorp/tfc-workflows-tooling/pull/57)
+
 ## Bug Fixes
 * Fixes issue with `payload` output missing from `run create`, `run show`, `upload`, `plan output` commands by @mjyocca [#46](https://github.com/hashicorp/tfc-workflows-tooling/pull/46)
 * Fixes race condition with the `run create` command when an `auto-apply` configured TFC/TFE workspace exits too soon by @mjyocca [#55](https://github.com/hashicorp/tfc-workflows-tooling/pull/55)